+++ src/components/common/Auth/index.tsx
@@ -1,0 +1,1 @@
+
